refactor(core): extract the shared AST toolkit into module-ast.ts and fold component-collect into component.ts
component-parse.ts was two modules under one name: the component-fact
parser, and a general AST toolkit (unwrapTs, the scoped walkers, the
program-level collectors, the suppression scanner) that the Kit-module and
config parsers imported by name. The toolkit now lives in module-ast.ts;
component-parse is one of its consumers like its siblings.

component-collect.ts was a shallow wrapper whose real content — the
empty-facts shape — had to be kept in sync with ComponentFacts through a
comment. emptyComponentFacts and collectComponentFacts now live next to
the shape in component.ts, where adding a field breaks the zero value in
the same file.

(cherry picked from commit d9ac7f1f363062edd7e17c81264fa4f61f0fe9d8)
