+++ packages/core/test/each-index-key.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { correctnessEachIndexKey } from '../src/rules/correctness/each-index-key.js';
 import { correctnessEachKey } from '../src/rules/correctness/each-key.js';
-import { emptyComponentFacts } from '../src/component-collect.js';
+import { emptyComponentFacts } from '../src/component.js';
 import { defaultProject, defineConfig } from '../src/types.js';
 import type { RuleContext } from '../src/rule.js';
 import type { ComponentFacts } from '../src/component.js';
